+++ src/main/index.ts
@@ -6,7 +6,7 @@ import * as fs from 'fs/promises';
 import * as os from 'os';
 import mainLogger from 'electron-log/main.js';
 
-import { initLogging, log } from './Log';
+import { initLogging, log } from './Logging';
 import { installExtension, REACT_DEVELOPER_TOOLS } from 'electron-devtools-installer';
 import { initializeSerialHandlers, cleanupSerialPorts } from './serialService';
 import { initializeSocketHandlers, cleanupSockets } from './socketService';